# api/wutsi-openapi/src/openapi/v2/marketplace-access.yaml
openapi: "3.0.0"
info:
  title: "Wutsi Marketplace Access API"
  description: Marketplace domain and persistence API
  version: "2.0.0"
servers:
  - url: https://marketplace-access-test.herokuapp.com
    description: Sandbox
  - url: https://marketplace-access-prod.herokuapp.com
    description: Production
paths:
...
  /v1/fundraisings/{id}/status:
    post:
      operationId: "update-fundraising-status"
      description: Update the status of the fundraising
      parameters:
        - in: path
          name: id
          description: ID of the fundraising
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFundraisingStatusRequest'
      responses:
        200:
          description: Success
      tags:
        - Store
...
components:
  schemas:
...
    UpdateFundraisingStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum:
            - UNDER_REVIEW
            - ACTIVE
            - INACTIVE
...